fest: + auth + rights

+++ controllers/cards.js
@@ -30,6 +30,7 @@ module.exports.createCard = (req, res) => {
 
 module.exports.deleteCard = (req, res) => {
   Card.findByIdAndRemove(req.params.cardId)
+    .where('owner').equals(req.user._id)
     .orFail(new Error('NoValidid'))
     .then((card) => res.status(200).send({ data: card }))
     .catch((err) => {
